pyproject.toml
[project]
name = "document-intelligence-refinery"
version = "0.1.0"
description = "Multi-stage agentic pipeline for document intelligence (TRP1 Week 3)."
requires-python = ">=3.10"
authors = [
  { name = "FDE Trainee", email = "you@example.com" }
]
readme = "README.md"

dependencies = [
  "pydantic>=2.7.0",
  "pdfplumber>=0.11.0",
  "docling>=2.4.0",
  "numpy>=1.26.0",
  "langgraph>=0.2.0",
  "chromadb>=0.5.0",
  "httpx>=0.27.0",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=5.0.0",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

